backend_v2/pyproject.toml
[project]
name = "qlue-ui-backend"
version = "0.1.0"
description = "Backend API for QLue-UI"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "pydantic>=2.12",
    "pyyaml>=6.0",
    "watchdog>=6.0.0",
]
